Added a repo layer to connect to DB

+++ Consumer/src/main/java/org/example/entity/Student.java
@@ -1,7 +1,6 @@
 package org.example.entity;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -16,7 +15,7 @@ public class Student {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     private String name;
